Adding the tie tree
Adding the tie tree and changing around the logic. There is no more auto extract, just parsing the absolute paths to the records and then moving to a Map<String, Record>. Next step would be moving from Map<String, Record> to a trie tree holding a reference to the record.